pyproject.toml
[project]
name = "daily-brief-agent"
version = "0.1.0"
description = "Daily Brief Agent"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Daily Brief Agent Contributors" }]
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "feedparser>=6.0.11",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.6.0",
]

[project.scripts]
daily-brief-agent = "daily_brief_agent.cli:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
addopts = "-q"
